+++ src/aperio-io.h
@@ -6,8 +6,8 @@ static const char SVS[] = "svs";
 // main functions
 int32_t is_aperio(const char *filename);
 
-int32_t handle_aperio(const char **filename, const char *new_label_name, bool keep_macro_image,
-                      bool disable_unlinking, bool do_inplace);
+int32_t handle_aperio(const char **filename, const char *new_label_name, bool keep_macro_image, bool disable_unlinking,
+                      bool do_inplace);
 
 // additional functions
 char *override_image_description(char *result, char *delimiter);